added demand parameter

+++ include/graph.h
@@ -19,7 +19,7 @@ class Graph{
     public:
         std::vector<Node> Nodes;
         int number_nodes;
-        std::vector<std::vector<float>> distances; //tablica z dlugosciami pomiedzy węzłami
+        std::vector<std::vector<double>> distances; //tablica z dlugosciami pomiedzy węzłami
         std::vector<Node> unvisited;
         std::vector<Truck> trucksvector;
 
@@ -34,7 +34,7 @@ class Graph{
         void create_trucks();
         void makeunvisitedvector();
         void show_one_node_values(Node& node);
-        float rungrasp();
+        double rungrasp();
         void reset_trucks();
         //...
